Disabled thread to core pinning, as it makes performance worse.
Note: If the remaining core indices do not go from 0 to nThreads - 1 the algorithm in init does not work

Disabled thread to core pinning, as it makes performance worse.

Note: If the remaining core indices do not go from 0 to nThreads - 1 the algorithm in init does not work

Disabled thread to core pinning, as it makes performance worse.

Note: If the remaining core indices do not go from 0 to nThreads - 1 the algorithm in init does not work